@@ -1,16 +1,35 @@
 {
   "name": "eyed",
-  "version": "3.2.0",
-  "main": "dist/main.cjs.js",
+  "version": "4.0.0",
+  "type": "module",
+  "main": "dist/main.cjs",
   "types": "dist/main.d.ts",
+  "exports": {
+    ".": {
+      "types": "./dist/main.d.ts",
+      "browser": {
+        "require": "./dist/main.cjs",
+        "default": "./dist/main.cjs",
+        "import": "./dist/main.mjs"
+      },
+      "default": {
+        "require": "./dist/main.cjs",
+        "import": "./dist/main.mjs",
+        "default": "./dist/main.mjs"
+      }
+    }
+  },
+  "files": [
+    "dist"
+  ],
   "scripts": {
-    "build": "webpack -c webpack.config.js",
+    "build": "node --loader ts-node/esm node_modules/webpack-cli/bin/cli.js --config webpack.config.ts",
     "prepublishOnly": "npm test && npm run build",
     "pretest": "npm run lint",
     "test": "npm run testonly",
-    "testonly": "mocha ./test",
+    "testonly": "npx tsx ./node_modules/.bin/mocha ./test",
     "lint": "eslint",
-    "upgrade": "npx npm-check --update"
+    "upgrade": "npx npm-check -u -i '\"{chai,mocha,@types/mocha,@types/chai}\"'"
   },
   "author": {
     "name": "Cory Mawhorter",
@@ -27,28 +46,28 @@
   "license": "MIT",
   "dependencies": {
     "base-x": "^5.0.1",
-    "uuid": "^10.0.0"
+    "uuid": "^13.0.0"
   },
   "devDependencies": {
-    "@eslint/js": "^9.38.0",
+    "@babel/preset-env": "^7.28.5",
+    "@babel/preset-typescript": "^7.28.5",
+    "babel-loader": "^10.0.0",
+    "@eslint/js": "^9.39.1",
     "@types/chai": "^4.3.16",
     "@types/mocha": "^10.0.7",
-    "@types/node": "^24.8.1",
-    "@types/uuid": "^10.0.0",
-    "chai": "^4.3.10",
-    "eslint": "^9.38.0",
-    "globals": "^16.4.0",
+    "@types/node": "^24.10.1",
+    "@types/uuid": "^11.0.0",
+    "chai": "4.x",
+    "eslint": "^9.39.1",
+    "globals": "^16.5.0",
     "mocha": "^10.5.2",
     "source-map-loader": "^5.0.0",
     "source-map-support": "^0.5.21",
     "ts-loader": "^9.5.4",
     "ts-node": "^10.9.2",
     "typescript": "^5.9.3",
-    "typescript-eslint": "^8.46.1",
+    "typescript-eslint": "^8.46.4",
     "webpack": "^5.102.1",
     "webpack-cli": "^6.0.1"
-  },
-  "files": [
-    "dist/"
-  ]
+  }
 }
